requirements.txt
Django==1.7.4
django-jinja==1.1.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
psycopg2==2.6
django-select2-forms==1.1.18
pytz==2014.10
django-modeltranslation==0.8
unicodecsv==0.9.4
django-grappelli==2.6.3
grappelli-modeltranslation==0.1.3
python-dateutil==2.4.1
